chore(tests): move ExtractRemainingFromTable to plugin_collection
