+++ falco_plugin_tests/src/plugin_collection/extract/mod.rs
@@ -1,0 +1,1 @@
+pub mod remaining_from_table;
