+++ falco_plugin_tests/src/plugin_collection/mod.rs
@@ -1,1 +1,2 @@
+pub mod extract;
 pub mod source;
