+++ falco_plugin_tests/tests/parse.rs
@@ -2,22 +2,12 @@ use falco_plugin::anyhow::Error;
 use falco_plugin::base::Plugin;
 use falco_plugin::event::events::types::EventType::PLUGINEVENT_E;
 use falco_plugin::event::events::types::{EventType, PPME_PLUGINEVENT_E};
-use falco_plugin::extract::{field, ExtractFieldInfo, ExtractPlugin, ExtractRequest};
 use falco_plugin::parse::{ParseInput, ParsePlugin};
 use falco_plugin::source::EventInput;
-use falco_plugin::tables::export;
-use falco_plugin::tables::import;
 use falco_plugin::tables::TablesInput;
 use falco_plugin::{anyhow, static_plugin};
+use falco_plugin_tests::plugin_collection::extract::remaining_from_table::RemainingEntryTable;
 use std::ffi::CStr;
-use std::sync::Arc;
-
-type RemainingEntryTable = export::Table<u64, RemainingCounter>;
-
-#[derive(export::Entry)]
-struct RemainingCounter {
-    remaining: export::Public<u64>,
-}
 
 struct ParseTestPlugin {
     remaining_table: Box<RemainingEntryTable>,
@@ -65,61 +55,12 @@ impl ParsePlugin for ParseTestPlugin {
     }
 }
 
-type RemainingCounterImportTable = import::Table<u64, RemainingCounterImport>;
-type RemainingCounterImport = import::Entry<Arc<RemainingCounterImportMetadata>>;
-
-#[derive(import::TableMetadata)]
-#[entry_type(RemainingCounterImport)]
-struct RemainingCounterImportMetadata {
-    remaining: import::Field<u64, RemainingCounterImport>,
-}
-
-struct DummyExtractPlugin {
-    remaining_table: RemainingCounterImportTable,
-}
-
-impl Plugin for DummyExtractPlugin {
-    const NAME: &'static CStr = c"dummy_extract";
-    const PLUGIN_VERSION: &'static CStr = c"0.0.0";
-    const DESCRIPTION: &'static CStr = c"test plugin";
-    const CONTACT: &'static CStr = c"rust@localdomain.pl";
-    type ConfigType = ();
-
-    fn new(input: Option<&TablesInput>, _config: Self::ConfigType) -> Result<Self, Error> {
-        let input = input.ok_or_else(|| anyhow::anyhow!("did not get table input"))?;
-        let remaining_table = input.get_table(c"remaining")?;
-
-        Ok(Self { remaining_table })
-    }
-}
-
-impl DummyExtractPlugin {
-    fn extract_remaining(&mut self, req: ExtractRequest<Self>) -> Result<u64, Error> {
-        let event_num = req.event.event_number() as u64;
-
-        let entry = self
-            .remaining_table
-            .get_entry(req.table_reader, &event_num)?;
-        let remaining = entry.get_remaining(req.table_reader)?;
-
-        Ok(remaining)
-    }
-}
-
-impl ExtractPlugin for DummyExtractPlugin {
-    const EVENT_TYPES: &'static [EventType] = &[PLUGINEVENT_E];
-    const EVENT_SOURCES: &'static [&'static str] = &["countdown"];
-    type ExtractContext = ();
-    const EXTRACT_FIELDS: &'static [ExtractFieldInfo<Self>] =
-        &[field("countdown.remaining", &Self::extract_remaining)];
-}
-
 static_plugin!(PARSE_PLUGIN_API = ParseTestPlugin);
-static_plugin!(DUMMY_EXTRACT_API = DummyExtractPlugin);
 
 #[cfg(test)]
 mod tests {
     use falco_plugin::base::Plugin;
+    use falco_plugin_tests::plugin_collection::extract::remaining_from_table::EXTRACT_REMAINING_FROM_TABLE_API;
     use falco_plugin_tests::plugin_collection::source::countdown::{
         CountdownPlugin, COUNTDOWN_PLUGIN_API,
     };
@@ -135,7 +76,7 @@ mod tests {
         .unwrap();
         let _ = driver.register_plugin(&super::PARSE_PLUGIN_API, c"");
         let extract_plugin = driver
-            .register_plugin(&super::DUMMY_EXTRACT_API, c"")
+            .register_plugin(&EXTRACT_REMAINING_FROM_TABLE_API, c"")
             .unwrap();
         driver
             .add_filterchecks(&extract_plugin, c"countdown")
